pyproject.toml
[tool.poetry]
name = "coursework-moshura"
version = "0.1.0"
description = ""
authors = ["GivenFLY <alexey.givenfly.moshura@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-filter = "^23.2"
django-dotenv = "^1.4.2"
django-cors-headers = "^4.1.0"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
